Map Python "int" type to PyLong. (GH-5830)
* On unary operators (+/-/~), Python builtin types usually return their own type or fail, so infer their type.

* Allow 'long' as type only in Py2 code (language_Level=2) but map it to 'int'.

Closes https://github.com/cython/cython/issues/4237

+++ docs/examples/userguide/fusedtypes/char_or_float.py
@@ -1,6 +1,4 @@
-from __future__ import print_function
-
-char_or_float = cython.fused_type(cython.char, cython.float)
+char_or_float = cython.fused_type(cython.char, cython.double)
 
 
 
@@ -12,6 +10,6 @@ def plus_one(var: char_or_float) -> char_or_float:
 def show_me():
 
     a: cython.char = 127
-    b: cython.float = 127
+    b: cython.double = 127
     print('char', plus_one(a))
     print('float', plus_one(b))
